% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age_event_cols.R
\name{add_age_event_cols}
\alias{add_age_event_cols}
\title{Adds column with age at the endpoint diagnosis}
\usage{
add_age_event_cols(study_data, endpt)
}
\arguments{
\item{study_data}{A data.frame with at least columns \code{DATE_OF_BIRTH},
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE} where
the columns are the study endpoint and date, which will
differ depending on the input variable \code{endpt}.}

\item{endpt}{A string. The current endpoint.}
}
\value{
The data.frame with tow added columns \code{AGE_AT_BASE} and \code{AGE_AT_END}
}
\description{
The age is calculated from birth and from baseline.
}
\author{
Kira E. Detrois
}
