% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age_obs_cols.R
\name{add_age_obs_cols}
\alias{add_age_obs_cols}
\title{Adds column with age at start and end of the observation period}
\usage{
add_age_obs_cols(study_data)
}
\arguments{
\item{study_data}{A data.frame with at least columns \code{DATE_OF_BIRTH},
\code{OBS_START_DATE} and \code{OBS_END_DATE}.}
}
\value{
The data.frame with tow added columns \code{AGE_AT_BASE} and \code{AGE_AT_END}
}
\description{
The age at the start of the observation period is the age at
baseline.
}
\author{
Kira E. Detrois
}
