% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_diag_time_cols.R
\name{add_diag_time_cols}
\alias{add_diag_time_cols}
\title{Adds columns for the date of onset and the age in days}
\usage{
add_diag_time_cols(pheno_data, study)
}
\arguments{
\item{pheno_data}{A data.frame with at least columns \code{DATE_OF_BIRTH}
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE}
where the columns are the study endpoint and
date, which will differ depending on the input
variable \code{endpt}.}

\item{study}{An S4 class representing the study setup.}
}
\value{
The data.frame with for example for endpoint \code{J10_ASTHMA},
added column \code{J10_ASTHMA_exact_age_yrs}, and changed column
\code{J10_ASTHMA_DATE} with the end of the study as the
date for controls.
}
\description{
Calculates the age at onset of the endpoint of interest.
The date of onset is either the date of diagnosis for cases or the
end of the study period for controls.
}
\author{
Kira E. Detrois
}
