% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_study_interval_cols.R
\name{add_study_interval_cols}
\alias{add_study_interval_cols}
\title{Adds columns for the study setup}
\usage{
add_study_interval_cols(pheno_data, study)
}
\arguments{
\item{pheno_data}{A data.frame with at least the column \code{DATE_OF_BIRTH}.}

\item{study}{An S4 class representing the study setup.}
}
\value{
The phenotype data.frame with added columns:
\itemize{
\item \code{ENDPT_FREE}: lubdridate intervals
which ranges from \code{DATE_OF_BIRTH} to the
end of the washout period. In this time
period there should be no endpoint
occurance for eligible study participants.
\item \code{STUDY_TIME}: lubdridate intervals which range from
the start of the exposure period to the
end of the prediction period.
\item \code{EXP_LEN}: Exact time length of the exposure period in
years.
}
}
\description{
Adds the three columns, \code{EXP_LEN}, \code{ENDPT_FREE}, and \code{STUDY_TIME}
to the phenotype data.frame. Columns \code{ENDPT_FREE}, and \code{STUDY_TIME}
each contain lubridate \code{\link[lubridate]{interval}}s.
}
\details{
The column \code{EXP_LEN} is mainly needed for studies that consider
variable exposure lengths for individuals selected at a given time point.
Here the exposure length will be the time from birth until the time point
considered and should be provided in the S4 study object in slot \code{exp_len}
in the same order as the individuals in the data.frame. A study setup like
this can be created using the function \code{\link{get_backward_study}}.
}
\author{
Kira E. Detrois
}
