% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\name{apply_endpt_filter}
\alias{apply_endpt_filter}
\title{Filters out cohorts that were used to train the}
\usage{
apply_endpt_filter(study_data, endpt)
}
\description{
Filters out cohorts that were used to train the
}
\author{
Kira E. Detrois
}
