% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_data_dates.R
\name{calc_backward_obs_start_date}
\alias{calc_backward_obs_start_date}
\title{Calcualtes the start date of the observation period for a backward study setup}
\usage{
calc_backward_obs_start_date(obs_end_date, obs_len)
}
\arguments{
\item{obs_len}{An integer. Length of the prediction period (in years).}

\item{obs_end_data}{A Date. The end of the observation period.}
}
\value{
A Date
}
\description{
Calcualtes the start date of the observation period for a backward study setup
}
\author{
Kira E. Detrois
}
