% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_diag_time_cols.R
\name{calc_diag_time}
\alias{calc_diag_time}
\title{Calculates the date of onset and the age in days}
\usage{
calc_diag_time(pheno_data, study)
}
\arguments{
\item{pheno_data}{A data.frame with at least columns \code{DATE_OF_BIRTH}
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE}
where the columns are the study endpoint and
date, which will differ depending on the input
variable \code{endpt}.}

\item{study}{An S4 class representing the study setup.}
}
\value{
A list(\code{exact_age_yrs}, \code{age_date}).
\itemize{
\item \code{exact_age_yrs}: The age at onset in days
\item \code{age_date}: The date of onset
}
}
\description{
Calculates the age at onset of the endpoint of interest.
The date of onset is either the date of diagnosis for cases or the
end of the study period for controls.
}
\author{
Kira E. Detrois
}
