% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_end_of_study.R
\name{calc_end_of_study}
\alias{calc_end_of_study}
\title{Calculates the end of the study time}
\usage{
calc_end_of_study(pheno_data, study)
}
\arguments{
\item{pheno_data}{A data.frame with at least columns \code{DATE_OF_BIRTH},
and \code{EXP_END}.}

\item{study}{An S4 class representing the study setup.}
}
\value{
A lubdridate interval. The total study time interval.
}
\description{
Calculates the end of the study time for each individual. The study time
is the time from the begining of the exposure period to the end
of the observation period.
}
\examples{
test_data <- Istudy::create_test_df(30)
study <- methods::new("study", endpt="J10_ASTHMA", exp_age=30, 
                      EXP_END=10, wash_len=2, obs_len=8, ancs="EUR")
test_data <- Istudy::add_study_interval_cols(test_data, study)
calc_end_of_study(test_data, study)

}
\author{
Kira E. Detrois
}
