% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endpt_free_time.R
\name{calc_endpt_free_time}
\alias{calc_endpt_free_time}
\title{Calcualtes endpoint free intervals}
\usage{
calc_endpt_free_time(pheno_data, study)
}
\arguments{
\item{pheno_data}{A data.frame with at least columns \code{DATE_OF_BIRTH},
and \code{EXP_END}.}

\item{study}{An S4 class representing the study setup.}

\item{bds}{A Date (vector). The birth days of the individuals.}
}
\value{
A lubdridate interval (vector). The endpoint free intervals.
}
\description{
Calculates the endpoint free interval for each individual. The
endpoint free interval is the period from birth until the
observation period begins. Thus, the interval contains the
exposure and washout period.
}
\details{
Creates lubridate \code{\link[lubridate]{interval}}s.
}
\author{
Kira E. Detrois
}
