% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_data_dates.R
\name{calc_exp_end_date}
\alias{calc_exp_end_date}
\title{Calcualtes the end date of the exposure period}
\usage{
calc_exp_end_date(study_data, exp_len)
}
\arguments{
\item{study_data}{A tibble. The data on all study individuals.}

\item{exp_len}{An integer. Length of the exposure period (in years).}
}
\value{
A vector of Dates.
}
\description{
Calcualtes the end date of the exposure period
}
\author{
Kira E. Detrois
}
