% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_exp_start_date.R
\name{calc_exp_start_date}
\alias{calc_exp_start_date}
\title{Calculates start of exposure period}
\usage{
calc_exp_start_date(bds, exp_age)
}
\arguments{
\item{bds}{A Date. The birth days of the individuals.}

\item{exp_age}{An integer. Age at which exposure period starts
(in years)}
}
\value{
A Date. The dates of the start of the exposure period
for each individual.
}
\description{
Calculates the date of the start of the exposure period for each
individual.
}
\examples{
bds <- c(as.Date("1923/07/01"), as.Date("1823/07/02"), as.Date("2002/04/01"))
study <- methods::new("study", endpt="J10_ASTHMA", exp_age=30, exp_len=10, wash_len=2, obs_len=8, ancs="EUR")
calc_exp_start_date(bds, study@exp_age)

}
\author{
Kira E. Detrois
}
