% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_data_dates.R
\name{calc_obs_end_date}
\alias{calc_obs_end_date}
\title{Calcualtes the start date of the observation period for each
individual}
\usage{
calc_obs_end_date(study_data, study_setup)
}
\arguments{
\item{study_data}{A tibble. The data on all study individuals.}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link{study_setup}.}
}
\value{
A vector of Dates. The start date of the observation period
for each individual.
}
\description{
Calcualtes the start date of the observation period for each
individual, depending on the selected study setup.
}
\author{
Kira E. Detrois
}
