% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_data_dates.R
\name{calc_obs_start_date}
\alias{calc_obs_start_date}
\title{Calcualtes the end date of the washout period for each individual}
\usage{
calc_obs_start_date(study_data, study_type, wash_len, obs_len, obs_end_date)
}
\arguments{
\item{study_data}{A tibble. The data on all study individuals.}

\item{study_type}{A character. Can be either \code{forward} or \code{backward}.}

\item{wash_len}{An integer. Length of the washout period (in years).}

\item{obs_len}{An integer. Length of the observation period
(in years).}

\item{obs_end_date}{A Date. The end of the observation period.}
}
\value{
A vector of Dates. The end of the washout period for
each individual.
}
\description{
Calcualtes the end date of the washout period for each individual,
depending on the selected study setup.
}
\author{
Kira E. Detrois
}
