% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_study_time.R
\name{calc_study_time}
\alias{calc_study_time}
\title{Calculates total study time interval}
\usage{
calc_study_time(pheno_data, study)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, \code{SEX}, \code{DATE_OF_BIRTH}, \code{ANCESTRY},
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP},
\code{DATE_OF_BIRTH}, and i.e. \code{J10_ASTHMA}, and
\code{J10_ASTHMA_DATE} where the columns are the study
endpoint and date, which will differ depending on
the input variable \code{endpt}.}

\item{study}{An S4 class representing the study setup.}
}
\value{
A lubdridate interval. The total study time interval.
}
\description{
Calculates the total study time for each individual. The study time
is the time from the begining of the exposure period to the end
of the prediction period.
}
\details{
Creates lubridate \code{\link[lubridate]{interval}}s.
}
\examples{
test_data <- Istudy::create_test_df(30)
study <- methods::new("study", endpt="J10_ASTHMA", exp_age=30, 
                      exp_len=10, wash_len=2, obs_len=8, ancs="EUR")
test_data <- Istudy::add_study_interval_cols(test_data, study)
calc_study_time(test_data, study)

}
\author{
Kira E. Detrois
}
