% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_data_dates.R
\name{calc_wash_end_date}
\alias{calc_wash_end_date}
\title{Calcualtes the end date of the washout period for each
individual}
\usage{
calc_wash_end_date(study_data, wash_len)
}
\arguments{
\item{study_data}{A tibble. The data on all study individuals.}

\item{wash_len}{An integer. Length of the washout period (in years).}
}
\value{
A vector of Dates.
}
\description{
Calcualtes the end date of the washout period for each
individual
}
\author{
Kira E. Detrois
}
