% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_data_dates.R
\name{calc_wash_start_date}
\alias{calc_wash_start_date}
\title{Calcualtes the end date of the exposure period for each individual}
\usage{
calc_wash_start_date(study_data, study_setup)
}
\arguments{
\item{study_data}{A tibble. The data on all study individuals.}
}
\value{
A vector of Dates. The end of the exposure period for
each individual.
}
\description{
Calcualtes the end date of the exposure period for each individual,
depending on the selected study setup.
}
\author{
Kira E. Detrois
}
