% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{check_res_dir}
\alias{check_res_dir}
\title{Checks whether to write a results file and if the directory exists}
\usage{
check_res_dir(write_res, res_dir)
}
\arguments{
\item{write_res}{A boolean. Defines whether to write the results to
a file or not. Default is FALSE.}

\item{res_dir}{A character. The directory to write the results to.}
}
\value{
A boolean. Whether to write the results to the \code{res_dir}.
}
\description{
If the directory does not exists tries to create it, recursively.
}
\author{
Kira E. Detrois
}
