% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_endpt_date_info.R
\name{complete_endpt_date_info}
\alias{complete_endpt_date_info}
\title{Adds endpoint date information for controls and censored individuals}
\usage{
complete_endpt_date_info(study_data, endpt)
}
\arguments{
\item{study_data}{A data.frame with at least columns \code{DATE_OF_BIRTH}
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE}
where the columns are the study endpoint and
date, which will differ depending on the input
variable \code{endpt}.}

\item{endpt}{A character. The column name of the current endpoint of
interest.}
}
\value{
The data.frame. For example for endpoint \code{J10_ASTHMA},
the data.frame will have and added column \code{J10_ASTHMA_AGE},
and for controls the column \code{J10_ASTHMA_DATE} now contains
the study end date.
}
\description{
For cases endpoint date is the date of diagnois.
For controls, it is the end of the observation period, and for
censored individuals it is the end of followup date.
}
\author{
Kira E. Detrois
}
