% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{create_return_tib}
\alias{create_return_tib}
\title{Creates a list with relevant information of the selected data}
\usage{
create_return_tib(pheno_data, endpt)
}
\arguments{
\item{pheno_data}{A data.frame with at least columns \code{STUDY_TIME}
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE}
where the columns are the study endpoint and
date, which will differ depending on the input
variable \code{endpt}.}

\item{endpt}{A character. The column name of the current endpoint of
interest.}
}
\value{
The results data.frame to return
}
\description{
Gets only the columns relevant for the current endpoint.
Creates a list with the data, as well as the selected
study period intervals.
}
\author{
Kira E. Detrois
}
