% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ancestry.R
\name{filter_ancestry}
\alias{filter_ancestry}
\title{Gets only the individuals from the specified ancestries}
\usage{
filter_ancestry(pheno_data, ancs = "EUR")
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
\code{ID}, \code{SEX}, \code{DATE_OF_BIRTH}, \code{ANCESTRY},
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP},
\code{DATE_OF_BIRTH}, and i.e. \code{J10_ASTHMA}, and
\code{J10_ASTHMA_DATE} where the columns are the study
endpoint and date, which will differ depending on
the input variable \code{endpt}.}

\item{ancs}{A character (vector). The ancestries to select.}
}
\value{
The filtered data.frame.
}
\description{
Gets only the individuals from the specified ancestries
}
\author{
Kira E. Detrois
}
