% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_early_endpt.R
\name{filter_early_endpt}
\alias{filter_early_endpt}
\title{Filters out individuals with early endpoints}
\usage{
filter_early_endpt(study_data, endpt)
}
\arguments{
\item{study_data}{A data.frame with at least columns \code{ENDPT_FREE_PERIOD}
and i.e. \code{J10_ASTHMA_DATE}, the study endpoint date,
which will differ depending on the input variable \code{endpt}.}

\item{endpt}{A character. The column name of the current endpoint of
interest.}
}
\value{
The filtered data.frame without individuals where the
endpoint occured before the prediction period.
}
\description{
Filters out individuals where the endpoint occured before the
observation period begins.
}
\details{
Note lubridate intervals are inclusive. So the end dates of the
endpoint free period are also included.
}
\author{
Kira E. Detrois
}
