% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_missing_endpt_data.R
\name{filter_missing_endpt_data}
\alias{filter_missing_endpt_data}
\title{Filters out missing information on the endpoint}
\usage{
filter_missing_endpt_data(pheno_data, endpt)
}
\arguments{
\item{pheno_data}{A data.frame with at least the columns:
i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE} where
the columns are the study endpoint and date,
which will differ depending on the input variable
\code{endpt}.}

\item{endpt}{A character. The column name of the current endpoint of
interest.}
}
\value{
The filtered data.frame without any missing information
on the endpoint of interest.
}
\description{
Filters out individuals with missing information about the
endpoint of interest, defined by variable \code{endpt}.
Missing information is defined by either no recorded endpoint
(endpt: NA), or for cases no recorded date of endpoint (endpt: 1,
endpt_date: NA).
}
\examples{
test_data <- create_test_df()
filter_missing_endpt_data(test_data, "J10_ASTHMA")

}
\author{
Kira E. Detrois
}
