% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indv_exp_len.R
\name{filter_not_all_na_cols}
\alias{filter_not_all_na_cols}
\title{Filters out columns from the data with only NA entries}
\usage{
filter_not_all_na_cols(pheno_data)
}
\arguments{
\item{pheno_data}{A data.frame.}
}
\description{
Filters out columns from the data with only NA entries
}
\author{
Kira E. Detrois
}
