% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_too_old_and_young.R
\name{filter_old_age}
\alias{filter_old_age}
\title{Censors individuals once they reach the maximum age}
\usage{
filter_old_age(study_data, max_obs_age)
}
\arguments{
\item{study_data}{A data.frame with at least column \code{EXP_END}.}

\item{max_obs_age}{A numeric. The maximum age an individuals should
be at baseline.}
}
\description{
Censors individuals once they reach the maximum age
}
\author{
Kira E. Detrois
}
