% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_too_old_and_young.R
\name{filter_too_old_and_young}
\alias{filter_too_old_and_young}
\title{Filters out individuals that are too old or young for the study}
\usage{
filter_too_old_and_young(study_data, study_type, max_age = 200)
}
\arguments{
\item{study_data}{A data.frame with at least column \code{EXP_END}.}

\item{study_type}{A character. Can be either \code{forward} or \code{backward}.}

\item{max_age}{A numeric. The maximum age for individuals in the study.
Individuals are censored, once they reach the maximum age.}
}
\description{
This is mainly needed for studies that consider variable exposure lengths for
individuals selected at a given time point.
Here the exposure length will be the time from birth until the time point
considered and should be provided in the S4 study object in slot \code{EXP_END}
in the same order as the individuals in the data.frame.
}
\details{
It might be that the birth of an individual is after the selected time point
so their exposure lengths are negative. These individuals are always removed.
}
\author{
Kira E. Detrois
}
