% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_backward_study.R
\name{get_backward_study}
\alias{get_backward_study}
\title{Creats and S4 study object from a set time point backwards}
\usage{
get_backward_study(
  pheno_data,
  endpt,
  EXP_END = NULL,
  wash_len = 2,
  obs_len = 8,
  obs_end = NULL,
  downsample_fctr = NA_real_,
  ancs = NA_character_
)
}
\arguments{
\item{pheno_data}{A data.frame with columns i.e. \code{J10_ASTHMA_DATE}, and
\code{I9_VTE_DATE}. The diagnosis dates for different
endpoints.}

\item{endpt}{A character. The column name of the current endpoint of
interest.}

\item{wash_len}{An integer. Length of the washout period (in years).}

\item{obs_len}{An integer. Legnth of the observation period (in years).}

\item{obs_end}{A Date. The end of the observation period.}

\item{downsample_fctr}{A numeric. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}

\item{ancs}{A character (vector). The ancestries to consider.}
}
\description{
Considers all individuals at a set time point. The observation and
washout period are calcualted backwards from this time point.
While the exposure period will be different for each individual
depending on their birth date.
}
\details{
If no end date is provided with input variable \code{obs_len} uses the
most recent date across the different \verb{_DATE} columns.
}
