% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ewo_file_name}
\alias{get_ewo_file_name}
\title{Creates file name describing the exposure, wash, and observation period.}
\usage{
get_ewo_file_name(study_type, exp_len, wash_len, obs_len)
}
\description{
Creates file name describing the exposure, wash, and observation period.
}
\author{
Kira E. Detrois
}
