% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indv_exp_len.R
\name{get_indv_exp_len}
\alias{get_indv_exp_len}
\title{Gets the exact exposure times in years of each individual in the data}
\usage{
get_indv_exp_len(pheno_data, wash_len = 2, obs_len = 8, obs_end = NULL)
}
\arguments{
\item{pheno_data}{A data.frame with columns i.e. \code{J10_ASTHMA_DATE}, and
\code{I9_VTE_DATE}. The diagnosis dates for different
endpoints.}

\item{wash_len}{An integer. Length of the washout period (in years).}

\item{obs_len}{An integer. Legnth of the observation period (in years).}

\item{obs_end}{A Date. The end of the observation period.}
}
\description{
Gets the exact exposure times in years of each individual in the data
}
\author{
Kira E. Detrois
}
