% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_indv_exp_len.R
\name{get_max_date}
\alias{get_max_date}
\title{Get's the most recent date across different \verb{_DATE} columns}
\usage{
get_max_date(pheno_data)
}
\arguments{
\item{pheno_data}{A data.frame with columns i.e. \code{J10_ASTHMA_DATE}, and
\code{I9_VTE_DATE}. The diagnosis dates for different
endpoints.}
}
\description{
Get's the most recent date across different \verb{_DATE} columns
}
\author{
Kira E. Detrois
}
