% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_n_cntrls}
\alias{get_n_cntrls}
\title{Counts the number of controls in the data}
\usage{
get_n_cntrls(pheno_data, endpt)
}
\arguments{
\item{pheno_data}{A data.frame with at least the column defined in \code{endpt}.}

\item{endpt}{A string. The column name of the current endpoint of
interest.}
}
\description{
Counts the number of controls in the data for a given endpoints. A
control is considered to be an individual with a zero in the respective
endpoint column. This does not consider things like the observation
period from the study setup. This should be handled prior using
functions \code{add_study_interval_cols}, and \code{adj_case_cntrl_status}.
}
\author{
Kira E. Detrois
}
