% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_elig_indv.R
\name{get_study_elig_indv}
\alias{get_study_elig_indv}
\title{Gets the individuals which are eligble given the study setup}
\usage{
get_study_elig_indv(study, max_age = 200, write_res = FALSE, res_dir = NULL)
}
\arguments{
\item{study}{An S4 class representing the study setup.}

\item{max_age}{A numeric. The maximum length of the exposure period.}

\item{write_res}{A boolean. Defines whether to write the results to
a file or not. Default is FALSE.}

\item{res_dir}{A character. The directory to write the results to.}

\item{study@study_data}{A data.frame with at least the columns:
\code{ID}, \code{SEX}, \code{DATE_OF_BIRTH}, \code{ANCESTRY},
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP},
\code{DATE_OF_BIRTH}, and i.e. \code{J10_ASTHMA}, and
\code{J10_ASTHMA_DATE} where the columns are the study
endpoint and date, which will differ depending on
the input variable \code{endpt}.}
}
\value{
A tibble with the information for the eligible individuals
with columns: \code{ID}, \code{SEX}, \code{DATE_OF_BIRTH},
\code{START_OF_FOLLOWUP}, \code{END_OF_FOLLOWUP}, \code{ANCESTRY},
and i.e. \code{J10_ASTHMA}, and \verb{J10_ASTHMA_DATE. where the last two columns are the study endpoint and date,  which will differ depending on the input variable }endpt`.
}
\description{
There are two types of study setup:
}
\details{
\itemize{
\item The first one considers individuals of a certain age
and a set exposure, washout, and observation periods
calcualted onwards from this age. It can simply be created
by creating a S4 study object.
\item The second one considers all individuals at a set time
point. The observation and washout period are calcualted
backwards from this time point. The exposure period will be
different for each individual depending on their birth date.
This setup can be created, using the function
\code{\link{get_backward_study}}.
}

\itemize{
\item The study setup consists of an exposure window,
a washout period, and a prediction period. See function
\code{\link{calc_study_time}}.
\item Eligible individuals cannot have missing data in the
column for the endpoint of interest. See function
\code{\link{filter_missing_endpt_data}}.
\item Eligible individuals cannot have the
selected endpoint of interest inside the endpoint free
period. The endpoint free interval is the period from birth
until the prediction period begins. See function
\code{\link{filter_early_endpt}}.
\item Individuals where the endpoint onset date is after the
observation has ended are considered controls in this setup.
}
}
\author{
Kira E. Detrois
}
