% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\name{initialize,study-method}
\alias{initialize,study-method}
\title{Constructor for the \code{study} class}
\usage{
\S4method{initialize}{study}(.Object, ...)
}
\description{
Constructor for the \code{study} class
}
