% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\name{initialize,study_setup-method}
\alias{initialize,study_setup-method}
\title{Constructor for the \code{study_setup} class}
\usage{
\S4method{initialize}{study_setup}(.Object, ...)
}
\description{
Constructor for the \code{study_setup} class
}
