% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\name{process_EDU}
\alias{process_EDU}
\title{Turn EDU codes into integers}
\usage{
process_EDU(study_data)
}
\arguments{
\item{study_data}{A tibble. The data on all study individuals. Needs
at least column \code{EDU}.}
}
\value{
The updated study data
}
\description{
Turn EDU codes into integers
}
\author{
Kira E. Detrois
}
