% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\name{process_ISCED_2011}
\alias{process_ISCED_2011}
\title{Turn ISCED_2011 codes into integers}
\usage{
process_ISCED_2011(study_data)
}
\arguments{
\item{study_data}{A tibble. The data on all study individuals. Needs
at least column \code{ISCED_2011}.}
}
\value{
The updated study data
}
\description{
Turn ISCED_2011 codes into integers
}
\author{
Kira E. Detrois
}
