% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_data_dates.R
\name{set_study_data_dates}
\alias{set_study_data_dates}
\title{Sets the important dates for the study individuals}
\usage{
set_study_data_dates(
  study_data,
  study_type = "forward",
  exp_age = NA_integer_,
  exp_len = NA_integer_,
  wash_len = 2,
  obs_len = 8,
  obs_end_date = as.Date("2021/01/01")
)
}
\arguments{
\item{study}{The S4 study object.}
}
\value{
The S4 study object with the \code{study_data} tibble updated
with the date columns.
}
\description{
Sets columns for \code{YEAR_OF_BIRTH}, \code{EXP_START_DATE}, \code{EXP_END_DATE},
\code{WASH_END_DATE}, \code{OBS_END_DATE}, \code{ENDPT_FREE_PERIOD}, \code{STUDY_TIME}.
For each individuals depending on the selected study setup.
}
\author{
Kira E. Detrois
}
