% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\name{setEndpt,study-method}
\alias{setEndpt,study-method}
\title{Sets \code{endpt} of the S4 study object to a new value}
\usage{
\S4method{setEndpt}{study}(.Object, endpt)
}
\arguments{
\item{.Object}{The S4 study object.}

\item{endpt}{A character. The new endpoint.}
}
\value{
The updated \code{study} object.
}
\description{
Sets \code{endpt} of the S4 study object to a new value
}
\author{
Kira E. Detrois
}
