% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\name{setEndpt}
\alias{setEndpt}
\title{Sets \code{endpt} of the S4 study object to a new value}
\usage{
setEndpt(.Object, endpt)
}
\arguments{
\item{.Object}{The S4 study object.}

\item{endpt}{A character. The new endpoint.}
}
\description{
Sets \code{endpt} of the S4 study object to a new value
}
