% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\docType{class}
\name{study-class}
\alias{study-class}
\alias{study}
\title{An S4 class representing the study setup}
\description{
An S4 class representing the study setup
}
\section{Slots}{

\describe{
\item{\code{study_setup}}{An S4 \code{study_setup} object. The current study setup.
See class definition \link{study_setup}.}

\item{\code{study_data}}{A tibble. The data on all study individuals.}

\item{\code{endpt}}{A character. The current endpoint of interest.}
}}

\author{
Kira E. Detrois
}
