% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\name{test_integer_correct}
\alias{test_integer_correct}
\title{Checks input integers}
\usage{
test_integer_correct(msg, var, var_name, not_na = FALSE)
}
\arguments{
\item{msg}{A string. Previous message to append the new message to.}

\item{var}{A numeric. The variable to check.}

\item{var_name}{A string. The variable name.}

\item{not_na}{A boolean. Whether the variable is allowed to be NA.}
}
\value{
The updated study data
}
\description{
Checks input integers
}
\author{
Kira E. Detrois
}
