% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\name{updateStudyData}
\alias{updateStudyData}
\title{Sets \code{endpt} of the S4 study object to a new value}
\usage{
updateStudyData(.Object, study_data)
}
\arguments{
\item{.Object}{The S4 study object.}

\item{endpt}{A character. The new endpoint.}
}
\description{
Sets \code{endpt} of the S4 study object to a new value
}
