% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_res_file}
\alias{write_res_file}
\title{Writes results to a tab-delim file}
\usage{
write_res_file(study)
}
\arguments{
\item{study}{An S4 study object. The current study setup.}
}
\description{
Writes results to a tab-delim file
}
\author{
Kira E. Detrois
}
