% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_res_files}
\alias{write_res_files}
\title{Writes results and log file}
\usage{
write_res_files(study)
}
\arguments{
\item{study}{An S4 study object. The current study setup.}
}
\description{
Prints to console if something goes wrong.
}
\author{
Kira E. Detrois
}
