% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_age_columns.R
\name{add_age_obs_cols}
\alias{add_age_obs_cols}
\title{Adds column with survival time from baseline}
\usage{
add_age_obs_cols(study_data)
}
\arguments{
\item{study_data}{A data.frame with at least columns \code{DATE_OF_BIRTH}
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE}
where the columns are the study endpoint and
date, which will differ depending on the input
variable \code{endpt}.}
}
\value{
The data.frame. For example for endpoint \code{J10_ASTHMA},
the data.frame will have updated column\code{J10_ASTHMA_DATE}.
}
\description{
Survival time is counted from the start of the observation
period until end of followup.
}
\author{
Kira E. Detrois
}
