% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diag_time.R
\name{add_diag_time_cols}
\alias{add_diag_time_cols}
\title{Adds columns for the date and exact age in years of onset}
\usage{
add_diag_time_cols(study_data, endpt)
}
\arguments{
\item{study_data}{A data.frame with at least columns \code{DATE_OF_BIRTH}
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE}
where the columns are the study endpoint and
date, which will differ depending on the input
variable \code{endpt}.}
}
\value{
The data.frame. For example for endpoint \code{J10_ASTHMA},
the data.frame will have and added column \code{J10_ASTHMA_AGE},
and for controls the column \code{J10_ASTHMA_DATE} now contains
the study end date.
}
\description{
Calculates the age at onset of the endpoint of interest.
For cases the date of onset is the date of diagnosis and for
controls it is the end of the study period.
}
\author{
Kira E. Detrois
}
