% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_diag_time_cols.R
\name{calc_diag_time}
\alias{calc_diag_time}
\title{Calculates the date and exact age in years of onset}
\usage{
calc_diag_time(pheno_data, study)
}
\arguments{
\item{pheno_data}{A data.frame with at least columns \code{DATE_OF_BIRTH}
and i.e. \code{J10_ASTHMA}, and \code{J10_ASTHMA_DATE}
where the columns are the study endpoint and
date, which will differ depending on the input
variable \code{endpt}.}

\item{study}{An S4 class representing the study setup.}
}
\value{
A list(\code{exact_age_yrs}, \code{age_date}).
\itemize{
\item \code{exact_age_yrs}: The age at onset in days
\item \code{age_date}: The date of onset
}
}
\description{
Calculates the age at onset of the endpoint of interest.
For cases the date of onset is the date of diagnosis and for
controls it is the end of the study period.
}
\author{
Kira E. Detrois
}
