% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_endpt_free_time.R
\name{calc_endpt_free_time}
\alias{calc_endpt_free_time}
\title{Calcualtes endpoint free intervals}
\usage{
calc_endpt_free_time(bds, study)
}
\arguments{
\item{bds}{A Date (vector). The birth days of the individuals.}

\item{study}{An S4 class representing the study setup.}
}
\value{
A lubdridate interval (vector). The endpoint free intervals.
}
\description{
Calculates the endpoint free interval for each individual. The
endpoint free interval is the period from birth until the
observation period begins. Thus, the interval contains the
exposure and washout period.
}
\details{
Creates lubridate \code{\link[lubridate]{interval}}s.
}
\examples{
bds <- c(as.Date("1923/07/01"), as.Date("1823/07/02"), as.Date("2002/04/01"))
study <- methods::new("study", endpt="J10_ASTHMA", exp_age=30, 
                      exp_len=10, wash_len=2, obs_len=8, ancs="EUR")
calc_endpt_free_time(bds, study)

}
\author{
Kira E. Detrois
}
