% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_data_dates.R
\name{calc_wash_end_date}
\alias{calc_wash_end_date}
\title{Calcualtes the end date of the washout period for each individual}
\usage{
calc_wash_end_date(.Object)
}
\arguments{
\item{.Object}{The S4 study object.}
}
\value{
A vector of Dates. The end of the washout period for
each individual.
}
\description{
Calcualtes the end date of the washout period for each individual,
depending on the selected study setup.
}
\author{
Kira E. Detrois
}
