% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample_cntrls.R
\name{downsample_cntrls}
\alias{downsample_cntrls}
\title{Downsamples the number of controls in the data}
\usage{
downsample_cntrls(pheno_data, study)
}
\arguments{
\item{pheno_data}{A data.frame with at least the column defined in \code{endpt}.}

\item{study}{An S4 class representing the study setup.}
}
\value{
The downsampled data.frame.
}
\description{
Given the same number of cases and the same indices
will always sample the same individuals as controls.
This allows reproducibility of the study selection.
}
\author{
Kira E. Detrois
}
