% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ancestry.R
\name{filter_ancestry}
\alias{filter_ancestry}
\title{Gets only the individuals from the specified ancestries}
\usage{
filter_ancestry(pheno_data, ancs = "EUR")
}
\arguments{
\item{ancs}{A character (vector). The ancestries to select.}
}
\value{
The filtered data.frame.
}
\description{
Gets only the individuals from the specified ancestries
}
\author{
Kira E. Detrois
}
