% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_early_endpt.R
\name{filter_early_endpt}
\alias{filter_early_endpt}
\title{Filters out individuals with early endpoints}
\usage{
filter_early_endpt(pheno_data, endpt)
}
\arguments{
\item{pheno_data}{A data.frame with at least columns \code{ENDPT_FREE}
and i.e. \code{J10_ASTHMA_DATE} where the columns are the study
endpoint and date, which will differ depending on the input
variable \code{endpt}.}

\item{endpt}{A character. The column name of the current endpoint of
interest.}
}
\value{
The filtered data.frame without individuals where the
endpoint occured before the prediction period.
}
\description{
Filters out individuals where the endpoint occured before the
observation period begins.
}
\details{
The endpoint free interval is the period from birth until the
observation period begins. Thus, the contains the exposure and
washout period. See function \code{\link{calc_endpt_free_time}}.

To create the endpoint free intervals, use either the
function \code{\link{calc_endpt_free_time}} directly or
\code{\link{add_study_interval_cols}}.
}
\examples{
test_data <- Istudy::create_test_df(30)
study <- methods::new("study", endpt="J10_ASTHMA", exp_age=30, 
                      exp_len=10, wash_len=2, obs_len=8, ancs="EUR")
test_data <- add_study_interval_cols(test_data, study)
filter_early_endpt(test_data, "J10_ASTHMA")
 
}
\author{
Kira E. Detrois
}
