% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_early_endpt.R
\name{filter_early_endpt}
\alias{filter_early_endpt}
\title{Filters out early endpoint}
\usage{
filter_early_endpt(pheno_data, endpt)
}
\arguments{
\item{pheno_data}{A data.frame with at least columns \code{ENDPT_FREE}
and i.e. \code{J10_ASTHMA_DATE} where the columns are the study
endpoint and date, which will differ depending on the input
variable \code{endpt}.}

\item{endpt}{A string. The column name of the current endpoint of
interest.}
}
\value{
The filtered data.frame without individuals where the
endpoint occured before the prediction period.
}
\description{
Filters out individuals where the endpoint occured before the
prediction period began.
}
\details{
The endpoint free interval is the period from birth until the
prediction period begins. Thus, the interval also contains the
exposure and washout period. See function
\code{\link{calc_endpt_free_time}}.

For the input data format see:
\href{https://docs.google.com/document/d/1GbZszpPeyf-hyb0V_YDx828YbM7woh8OBJhvzkEwo2g/edit}{INTERVENE Phenotype File Definition}. To create the endpoint free interval, use either the
function \code{\link{calc_endpt_free_time}} directly or
\code{\link{add_study_interval_cols}}.
}
\examples{
test_data <- Istudy::create_test_df(30)
study <- methods::new("study", endpt="J10_ASTHMA", exp_age=30, exp_len=10, wash_len=2, obs_len=8)
test_data <- add_study_interval_cols(test_data, study)
filter_early_endpt(test_data, "J10_ASTHMA")
 
}
\author{
Kira E. Detrois
}
