% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_too_old_and_young.R
\name{filter_old_age}
\alias{filter_old_age}
\title{Censors individuals once they reach the maximum age}
\usage{
filter_old_age(study_data, max_age = 200)
}
\arguments{
\item{study_data}{A data.frame with at least column \code{EXP_END}.}

\item{max_age}{A numeric. The maximum age for individuals in the study.
Individuals are censored, once they reach the maximum age.}
}
\description{
Censors individuals once they reach the maximum age
}
