% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_too_old_and_young.R
\name{filter_too_old_and_young}
\alias{filter_too_old_and_young}
\title{Filters out individuals that are too old or young for the study}
\usage{
filter_too_old_and_young(pheno_data, max_age = 90)
}
\arguments{
\item{pheno_data}{A data.frame with at least column \code{EXP_LEN}.}

\item{max_age}{A numeric. The maximum length of the exposure period.}
}
\description{
This is mainly needed for studies that consider variable exposure lengths for
individuals selected at a given time point.
Here the exposure length will be the time from birth until the time point
considered and should be provided in the S4 study object in slot \code{exp_len}
in the same order as the individuals in the data.frame. A study setup like
this can be created using the function \code{\link{get_backward_study}}.
}
\details{
It might be that the birth of an individual is after the selected time point
so their exposure lengths are negative. These individuals are always removed.
Additionally, individuals are filtered out based on some maximum exposure
length which can be set using the input variable \code{max_age}.
}
\author{
Kira E. Detrois
}
