% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_study_elig_indv.R
\name{get_study_elig_indv}
\alias{get_study_elig_indv}
\title{Gets the individuals which are eligble given the study setup}
\usage{
get_study_elig_indv(study)
}
\arguments{
\item{study}{An S4 object with the current study setup.}
}
\description{
There are two types of study setup:
}
\details{
\itemize{
\item \code{forward} considers individuals of a certain age.
It calculates the exposure, washout, and observation period
onwards from this age.
\item \code{backward} considers all individuals at a set time
point. The observation and washout period are calcualted
backwards from this time point.
}

\itemize{
\item The study setup consists of an exposure period,
a washout period, and a observation period.
\item Eligible individuals cannot have missing data in the
column for the endpoint of interest. See function
\code{\link{filter_missing_endpt_data}}.
\item Eligible individuals cannot have the endpoint of interest
inside the endpoint free period. The endpoint free interval
ranges from birth until the observation period begins.
See function \code{\link{filter_early_endpt}}.
\item Individuals where the endpoint diagnosis date is after the
study has ended, are considered controls.
}
}
\author{
Kira E. Detrois
}
