% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_study_data_dates.R
\name{set_study_dates}
\alias{set_study_dates}
\title{Sets the important dates for the study individuals}
\usage{
set_study_dates(study_data, study_setup)
}
\arguments{
\item{study_data}{A tibble. The data on all study individuals.}

\item{study_setup}{An S4 \code{study_setup} object. The current study setup.
See class definition \link{study_setup}.}
}
\description{
Sets columns for \code{YEAR_OF_BIRTH}, \code{EXP_START_DATE}, \code{EXP_END_DATE},
\code{WASH_END_DATE}, \code{OBS_END_DATE}, \code{ENDPT_FREE_PERIOD}, \code{STUDY_TIME}.
For each individuals depending on the selected study setup.
}
\author{
Kira E. Detrois
}
