% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\docType{class}
\name{study-class}
\alias{study-class}
\alias{study}
\title{An S4 class representing the study setup}
\description{
An S4 class representing the study setup
}
\section{Slots}{

\describe{
\item{\code{study_type}}{A character. Can be either \code{forward} or \code{backward}.
\code{forward} considers individuals of a certain age.
It calculates the exposure, washout, and observation
period onwards from this age.
\code{backward} considers all individuals at a set time
point. The observation and washout period are calcualted
backwards from this time point.}

\item{\code{study_data}}{A tibble. The data on all study individuals.}

\item{\code{endpt}}{A character. The column name of the current endpoint of
interest.}
}}

\author{
Kira E. Detrois
}
