% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\docType{class}
\name{study-class}
\alias{study-class}
\alias{study}
\title{An S4 class representing the study setup}
\description{
An S4 class representing the study setup
}
\section{Slots}{

\describe{
\item{\code{study_type}}{A character. Can be either \code{forward} or \code{backward}.
\code{forward} considers individuals of a certain age
and a set exposure, washout, and observation periods
calcualted onwards from this age. It can simply be created
by creating a S4 study object.
\code{backward} considers all individuals at a set time
point. The observation and washout period are calcualted
backwards from this time point. The exposure period will be
different for each individual depending on their birth date.
This setup can be created, using the function
\code{\link{get_backward_study}}.}

\item{\code{endpt}}{A character. The column name of the current endpoint of interest.}

\item{\code{exp_age}}{An integer. Age at which exposure period starts (in years).}

\item{\code{exp_len}}{An integer. Length of the exposure period (in years).}

\item{\code{exp_ids}}{A character (vector). The IDs for the exposure lengths,
if they differ between individuals.}

\item{\code{wash_len}}{An integer. Length of the washout period (in years).}

\item{\code{obs_len}}{An integer. Length of the observation period (in years).}

\item{\code{obs_end}}{A Date. The end of the observation period.}

\item{\code{downsample_fctr}}{An integer. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}

\item{\code{ancs}}{A character (vector). The ancestries to consider.}
}}

\author{
Kira E. Detrois
}
