% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study_class_def.R
\docType{class}
\name{study-class}
\alias{study-class}
\alias{study}
\title{An S4 class representing the study setup}
\description{
An S4 class representing the study setup
}
\section{Slots}{

\describe{
\item{\code{endpt}}{A character. The column name of the current endpoint of interest.}

\item{\code{exp_age}}{An integer. Age at which exposure period starts (in years).}

\item{\code{exp_len}}{An integer. Length of the exposure period (in years).}

\item{\code{wash_len}}{An integer. Length of the washout period (in years).}

\item{\code{obs_len}}{An integer. Length of the observation period (in years).}

\item{\code{downsample_fctr}}{An integer. Defines how many controls there
should be for every case.
Default is NA, which means no
downsampling is performed.}
}}

\author{
Kira E. Detrois
}
