% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_output.R
\name{write_log_file}
\alias{write_log_file}
\title{Writes study setup to a file}
\usage{
write_log_file(study)
}
\arguments{
\item{study}{An S4 study object. The current study setup.}
}
\description{
Writes study setup to a file
}
\author{
Kira E. Detrois
}
